Revert "frontend: [onnx] unify type builders (#2334)"
This reverts commit 81b3c680932e284abfe1c9278b434d7ba14da3d9.
