+++ tests/frontend/onnx/test_tensor_type.py
@@ -1,52 +1,14 @@
-import onnx
 import pytest
 
-from xdsl.dialects.builtin import Float32Type, TensorType, f32, f64
+from xdsl.dialects.builtin import TensorType, f32, f64
 
 try:
     from onnx import TensorShapeProto, TypeProto  # noqa: E402
 
-    from xdsl.frontend.onnx.type import (
-        get_elem_type,  # noqa: E402
-        get_shape,  # noqa: E402
-        get_tensor_type,  # noqa: E402
-        get_type,  # noqa: E402
-    )
-    from xdsl.utils.hints import isa
+    from xdsl.frontend.onnx.shape_type import get_shape, get_tensor_type  # noqa: E402
 except ImportError as exc:
     print(exc)
     pytest.skip("onnx is an optional dependency", allow_module_level=True)
-
-
-def test_get_elem_type():
-    # test case 1: check if 1 corresponds to f32
-    assert get_elem_type(1) == f32
-
-    # test case 11: check if 11 corresponds to f64
-    assert get_elem_type(11) == f64
-
-    # test case -1: check if -1 (or other illegal values) corresponds to None
-    with pytest.raises(ValueError, match="Unknown elem_type: -1"):
-        get_elem_type(-1)
-
-
-def test_get_type():
-    tensor_type = onnx.TypeProto()
-    tensor_type.tensor_type.elem_type = onnx.TensorProto.FLOAT
-    tensor_type.tensor_type.shape.dim.extend(
-        [
-            onnx.TensorShapeProto.Dimension(dim_value=3),
-            onnx.TensorShapeProto.Dimension(dim_value=4),
-            onnx.TensorShapeProto.Dimension(dim_value=5),
-        ]
-    )
-    tt = get_type(tensor_type)
-
-    assert isa(tt, TensorType[Float32Type])
-
-    assert tt.get_num_dims() == 3
-    assert tt.get_shape() == (3, 4, 5)
-    assert tt.get_element_type().name == "f32"
 
 
 def test_get_shape():
